requirements.txt
streamlit 
requests
langchain
langgrapgh
openai-whisper
setuptools-rust
audio-recorder-streamlit
python-multipart
twilio